+++ ajax.php
@@ -24,16 +24,10 @@ function login(){
     $success = false;
     if (!empty($_POST)) {
         if (!empty($_POST['email']) && !empty($_POST['password'])) {
-
             $email = filter_var($_POST['email'], FILTER_VALIDATE_EMAIL);
             $account = Account::getAccountByEmail($email);
-
-            // Verify account password and set $_SESSION
             if ($account !== false) {
-                // echo '<pre>'.var_export(print_r($account, true)).'</pre>';
-                $pHash = password_hash($account->password, 1);
-                if (password_verify($_POST['password'], $pHash)) {
-//                    $_SESSION['account_id'] = $account->accountId;
+                if (password_verify($_POST['password'], $account->password)) {
                     $success = true;
                 }
             }
