+++ react/src/Components/main.js
@@ -3,7 +3,7 @@ import { BrowserRouter, Switch, Route } from "react-router-dom";
 import landingPage from "./landingPage";
 import getStarted from "./getStarted";
 import tempPage from "../tempPage";
-import loginpage from "../loginpage";
+import loginpage from "./loginpage";
 
 const Main = () => (
   <BrowserRouter>
